requirement.txt
streamlit  
pandas 
seaborn 
numpy 
matplotlib.pyplot